cmd_runner_fmt: refactor out to its own file
